separated components in separate folders
